Move sync types to the eth-rpc
