+++ substrate/frame/revive/rpc/src/types/sync.rs
@@ -1,0 +1,51 @@
+// This file is part of Substrate.
+
+// Copyright (C) Parity Technologies (UK) Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::*;
+use serde::{Deserialize, Serialize};
+
+/// Syncing status
+#[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(untagged)]
+pub enum SyncingStatus {
+	/// Syncing progress
+	SyncingProgress(SyncingProgress),
+	/// Not syncing
+	/// Should always return false if not syncing.
+	Bool(bool),
+}
+
+impl Default for SyncingStatus {
+	fn default() -> Self {
+		SyncingStatus::SyncingProgress(Default::default())
+	}
+}
+
+/// Syncing progress
+#[derive(Debug, Default, Clone, Serialize, Deserialize, Eq, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct SyncingProgress {
+	/// Current block
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub current_block: Option<U256>,
+	/// Highest block
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub highest_block: Option<U256>,
+	/// Starting block
+	#[serde(skip_serializing_if = "Option::is_none")]
+	pub starting_block: Option<U256>,
+}
