+++ substrate/frame/revive/src/evm/api/rpc_types_gen.rs
@@ -192,22 +192,6 @@ pub struct ReceiptInfo {
 	pub r#type: Option<Byte>,
 }
 
-/// Syncing status
-#[derive(Debug, Clone, Serialize, Deserialize, From, TryInto, Eq, PartialEq)]
-#[serde(untagged)]
-pub enum SyncingStatus {
-	/// Syncing progress
-	SyncingProgress(SyncingProgress),
-	/// Not syncing
-	/// Should always return false if not syncing.
-	Bool(bool),
-}
-impl Default for SyncingStatus {
-	fn default() -> Self {
-		SyncingStatus::SyncingProgress(Default::default())
-	}
-}
-
 /// Address(es)
 #[derive(Debug, Clone, Serialize, Deserialize, From, TryInto, Eq, PartialEq)]
 #[serde(untagged)]
@@ -275,21 +259,6 @@ pub struct Log {
 	pub transaction_hash: H256,
 	/// transaction index
 	pub transaction_index: U256,
-}
-
-/// Syncing progress
-#[derive(Debug, Default, Clone, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct SyncingProgress {
-	/// Current block
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub current_block: Option<U256>,
-	/// Highest block
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub highest_block: Option<U256>,
-	/// Starting block
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub starting_block: Option<U256>,
 }
 
 /// Filter Topic List Entry
